Remove specialized getter from the ParameterObserver interface.
* Fix meta examples;
* Fix static casting issue when unsubscribing an observer;
* Refactor all CrossValidationStorage and observers to use only CEvaluationResult
  as base class;

+++ src/shogun/base/SGObject.cpp
@@ -829,7 +829,7 @@ void CSGObject::unsubscribe(ParameterObserver *obs) {
 	it->second.unsubscribe();
 	m_subscriptions.erase(index);
 
-	obs->put("subscription_id", -1);
+	obs->put("subscription_id", static_cast<int64_t>(-1));
 }
 
 void CSGObject::observe(const Some<ObservedValue> value) const
